+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/page/workspaces/summary/AclEditor.scala
@@ -1,9 +1,9 @@
 package org.broadinstitute.dsde.firecloud.page.workspaces.summary
 
-import org.broadinstitute.dsde.firecloud.api.WorkspaceAccessLevel.WorkspaceAccessLevel
 import org.broadinstitute.dsde.firecloud.component._
 import org.broadinstitute.dsde.firecloud.component.Component._
 import org.broadinstitute.dsde.firecloud.page.OKCancelModal
+import org.broadinstitute.dsde.workbench.service.WorkspaceAccessLevel.WorkspaceAccessLevel
 import org.openqa.selenium.WebDriver
 
 /**
