+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/test/api/orch/OrchestrationApiSpec.scala
@@ -3,15 +3,15 @@ package org.broadinstitute.dsde.firecloud.test.api.orch
 import akka.http.scaladsl.model.StatusCodes
 import com.google.api.client.googleapis.json.GoogleJsonResponseException
 import com.google.api.services.bigquery.model.GetQueryResultsResponse
-import org.broadinstitute.dsde.firecloud.api.{APIException, Orchestration}
+import org.broadinstitute.dsde.firecloud.fixture.BillingFixtures
 import org.broadinstitute.dsde.workbench.auth.AuthToken
 import org.broadinstitute.dsde.workbench.config.{Credentials, UserPool}
-import org.broadinstitute.dsde.firecloud.fixture.BillingFixtures
+import org.broadinstitute.dsde.workbench.dao.Google.googleBigQueryDAO
+import org.broadinstitute.dsde.workbench.model.google.GoogleProject
+import org.broadinstitute.dsde.workbench.service.{APIException, Orchestration}
 import org.scalatest.concurrent.{Eventually, ScalaFutures}
 import org.scalatest.time.{Seconds, Span}
 import org.scalatest.{FreeSpec, Matchers}
-import org.broadinstitute.dsde.firecloud.dao.Google.googleBigQueryDAO
-import org.broadinstitute.dsde.workbench.model.google.GoogleProject
 
 class OrchestrationApiSpec extends FreeSpec with Matchers with ScalaFutures with Eventually
   with BillingFixtures {
