+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/test/e2e/SmoketestSpec.scala
@@ -2,14 +2,15 @@ package org.broadinstitute.dsde.firecloud.test.e2e
 
 import java.util.UUID
 
-import org.broadinstitute.dsde.workbench.config.Config
-import org.broadinstitute.dsde.firecloud.fixture.{SimpleMethodConfig, TestData, UserFixtures, WorkspaceFixtures}
+import org.broadinstitute.dsde.firecloud.fixture.{TestData, UserFixtures}
 import org.broadinstitute.dsde.firecloud.page.workspaces.methodconfigs.WorkspaceMethodConfigListPage
 import org.broadinstitute.dsde.firecloud.page.workspaces.WorkspaceDataPage
 import org.broadinstitute.dsde.firecloud.page.workspaces.summary.WorkspaceSummaryPage
-import org.broadinstitute.dsde.firecloud.test.WebBrowserSpec
-import org.scalatest._
 import org.broadinstitute.dsde.firecloud.test.Tags
+import org.broadinstitute.dsde.workbench.config.Config
+import org.broadinstitute.dsde.workbench.fixture.{SimpleMethodConfig, WorkspaceFixtures}
+import org.broadinstitute.dsde.workbench.service.test.WebBrowserSpec
+import org.scalatest._
 
 class SmoketestSpec extends FreeSpec with WebBrowserSpec
   with UserFixtures with WorkspaceFixtures with Matchers {
